✅ Update log models to pass new test files

+++ src/modules/logs/content/log_event_content/fetch_remote_module_event.rs
@@ -17,5 +17,5 @@ pub struct FetchRemoteModuleEvent {
     pub ship: ShipType,
 
     #[serde(rename = "ShipID")]
-    pub ship_id: u8,
+    pub ship_id: u64,
 }
