+++ src/modules/logs/content/log_event_content/module_store_event.rs
@@ -17,6 +17,6 @@ pub struct ModuleStoreEvent {
     pub ship: ShipType,
 
     #[serde(rename = "ShipID")]
-    pub ship_id: u8,
+    pub ship_id: u64,
     pub hot: bool,
 }
