+++ src/modules/logs/content/log_event_content/shipyard_new_event.rs
@@ -11,5 +11,5 @@ pub struct ShipyardNewEvent {
     pub ship_type_localized: Option<String>,
 
     #[serde(rename = "NewShipID")]
-    pub new_ship_id: u8,
+    pub new_ship_id: u64,
 }
